feat: finish authorization method

+++ src/main/java/TO/project/CinemaStreet/service/UserService.java
@@ -1,8 +1,11 @@
 package TO.project.CinemaStreet.service;
 
+import TO.project.CinemaStreet.CurrentUserContext;
+import TO.project.CinemaStreet.Permissions;
 import TO.project.CinemaStreet.Roles;
 import TO.project.CinemaStreet.model.User;
 import TO.project.CinemaStreet.utils.AuthenticationResponse;
+import TO.project.CinemaStreet.utils.NoPermissionException;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import org.springframework.stereotype.Service;
@@ -12,6 +15,8 @@ import java.util.List;
 @Service
 public class UserService{
     private final UserRepository userRepository;
+    private CurrentUserContext currentUserContext = new CurrentUserContext();
+
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
@@ -19,9 +24,9 @@ public class UserService{
     public AuthenticationResponse authenticate(String username, String password){
         if (userRepository.existsByUsernameAndPassword(username, password)) {
             User user = userRepository.getReferenceByUsernameAndPassword(username, password);
-            return new AuthenticationResponse(user.getId(), user.getRole(), true);
+            return new AuthenticationResponse(user.getId(), user.getRole(), true, user.getUsername(), user.getFirstName(), user.getLastName());
         }
-        return new AuthenticationResponse(0, Roles.EMPLOYEE, false);
+        return new AuthenticationResponse(0, Roles.EMPLOYEE, false, "", "", "");
     }
 
     public List<User> getAllUsers()
@@ -29,18 +34,25 @@ public class UserService{
         return userRepository.findAll();
 //        return FXCollections.observableArrayList(new User("Jan", "Kowalski", "email@gmail.com", Role.USER));
     }
-    public void addUser(User user)
-    {
-        userRepository.save(user);
+    public void addUser(User user) throws NoPermissionException {
+        if (currentUserContext.getRole().authorize(Permissions.ADD_USER)) {
+            userRepository.save(user);
+        } else {
+            throw new NoPermissionException("No permission");
+        }
     }
 
-    public boolean deleteUserById(Integer id){
-        if(userRepository.existsById(id))
-        {
-            userRepository.deleteById(id);
-            return true;
+    public boolean deleteUserById(Integer id) throws NoPermissionException {
+        if (currentUserContext.getRole().authorize(Permissions.REMOVE_USER)) {
+            if(userRepository.existsById(id))
+            {
+                userRepository.deleteById(id);
+                return true;
+            }
+            return false;
+        } else {
+            throw new NoPermissionException("No permission");
         }
 
-        return false;
     }
 }
